[fix] remove dtype=int

+++ supervised_learning/0x01-multiclass_classification/0-one_hot_encode.py
@@ -18,6 +18,6 @@ def one_hot_encode(Y, classes):
     if Y.size is 0:
         return None
     else:
-        ohe = np.zeros((classes, Y.shape[0]), dtype=int)
+        ohe = np.zeros((classes, Y.shape[0]))
         ohe[Y, np.arange(Y.size)] = 1
         return ohe
